feat: validate order in payment

+++ src/core/domain/entity/order.go
@@ -10,7 +10,7 @@ type Order struct {
 	ID          int              `json:"id" gorm:"primaryKey;autoIncrement"`
 	ClientId    int              `json:"client_id"`
 	StatusOrder enum.StatusOrder `json:"status_order"`
-	Amount      float64          `json:"amount"`
+	Amount      float32          `json:"amount"`
 	CreatedAt   time.Time        `json:"created_at,omitempty"`
 	Products    []*Product       `json:"products" gorm:"many2many:orders_products"`
 }
